requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.3  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.3.0  # https://github.com/jazzband/django-redis
django-rest-auth==0.9.5
dj-rest-auth==4.0.1
drf-spectacular==0.26.4
alabaster==0.7.13
# argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
black==23.3.0
boto3==1.28.18
botocore==1.31.18
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
# crispy-bootstrap5==0.7
cryptography==41.0.3
cssbeautifier==1.14.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
# Django==4.2.3
# django-allauth==0.54.0
django-anymail==10.0
django-coverage-plugin==3.0.0
# django-crispy-forms==2.0
django-debug-toolbar==4.1.0
# django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
# django-model-utils==4.3.1
# django-redis==5.3.0
django-storages==1.13.2
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.31.1
docutils==0.19
drf-yasg==1.21.7
EditorConfig==0.12.3
executing==1.2.0
factory-boy==3.2.1
Faker==19.2.0
filelock==3.12.2
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
# hiredis==2.2.3
html-tag-names==0.1.2
html-void-elements==0.1.0
identify==2.5.26
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
livereload==2.6.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
# Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg2==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.4.0
pytest-django==4.5.2
pytest-sugar==0.9.7
python-dateutil==2.8.2
# python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
# redis==4.6.0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.6.2
termcolor==2.3.0
text-unidecode==1.3
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
# whitenoise==6.5.0
wrapt==1.15.0
gunicorn==20.1.0

 